api/e2e/requirements.txt
loguru
minio
black
pymongo
pytest
requests
python-dotenv
